requirements.txt
anyio==4.12.1
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
cssselect2==0.9.0
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.0
idna==3.11
Jinja2==3.1.6
latex_formatter==1.2.7
lxml==6.0.2
Markdown==3.10.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
pycparser==2.23
pydyf==0.12.1
PyMuPDF==1.26.5
pypandoc==1.16.2
pypdfium2==5.3.0
pyphen==0.17.2
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
PyYAML==6.0.3
regex==2026.1.15
safetensors==0.7.0
setuptools==80.10.2
shellingham==1.5.4
sympy==1.14.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.1.0
typer-slim==0.21.1
typing_extensions==4.15.0
weasyprint==68.1
webencodings==0.5.1
xlsxwriter==3.2.9
zopfli==0.4.1
